+++ common/src/main/java/me/rufia/fightorflight/entity/ai/tasks/FOFPokemonAttackTask.java
@@ -4,10 +4,13 @@ import com.cobblemon.mod.common.battles.BattleRegistry;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import me.rufia.fightorflight.PokemonInterface;
 import me.rufia.fightorflight.entity.PokemonAttackEffect;
+import me.rufia.fightorflight.utils.PokemonUtils;
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.memory.MemoryModuleType;
 
 public abstract class FOFPokemonAttackTask {
-     public static int getAttackTime(PokemonEntity pokemonEntity) {
+    public static int getAttackTime(PokemonEntity pokemonEntity) {
         return ((PokemonInterface) pokemonEntity).getAttackTime();
     }
 
@@ -16,9 +19,13 @@ public abstract class FOFPokemonAttackTask {
     }
 
     public static boolean isTargetInBattle(PokemonEntity pokemonEntity) {
-        if (pokemonEntity.getTarget() instanceof ServerPlayer targetAsPlayer) {
+        if (getTarget(pokemonEntity) instanceof ServerPlayer targetAsPlayer) {
             return BattleRegistry.getBattleByParticipatingPlayer(targetAsPlayer) != null;
         }
         return false;
+    }
+
+    public static LivingEntity getTarget(PokemonEntity pokemonEntity) {
+        return PokemonUtils.getTarget(pokemonEntity);
     }
 }