+++ common/src/main/java/me/rufia/fightorflight/entity/ai/tasks/FOFPokemonMeleeTask.java
@@ -26,14 +26,17 @@ public class FOFPokemonMeleeTask {
                 ).apply(context, (lookTargetAccessor, attackTargetAccessor, attackCooldownAccessor, visibleMobsAccessor) -> ((serverLevel, livingEntity, l) -> {
                     LivingEntity target = context.get(attackTargetAccessor);
                     if (livingEntity instanceof PokemonEntity pokemonEntity) {
-                        if (canPerformAttack(pokemonEntity, target)) {
-                            lookTargetAccessor.set(new EntityTracker(target, true));
-                            FOFPokemonAttackTask.resetAttackTime(pokemonEntity, 0);
-                            pokemonEntity.swing(InteractionHand.MAIN_HAND);
-                            pokemonDoHurtTarget(pokemonEntity, target);
-                            ((PokemonInterface) pokemonEntity).setAttackTime(cooldownBetweenAttacks);
-                            attackCooldownAccessor.setWithExpiry(true, cooldownBetweenAttacks);
-                            return true;
+                        if (PokemonUtils.shouldMelee(pokemonEntity)) {
+                            if (canPerformAttack(pokemonEntity, target)) {
+                                lookTargetAccessor.set(new EntityTracker(target, true));
+                                FOFPokemonAttackTask.resetAttackTime(pokemonEntity, 0);
+                                pokemonEntity.swing(InteractionHand.MAIN_HAND);
+                                pokemonDoHurtTarget(pokemonEntity, target);
+                                pokemonEntity.setTarget(target);
+                                ((PokemonInterface) pokemonEntity).setAttackTime(cooldownBetweenAttacks);
+                                attackCooldownAccessor.setWithExpiry(true, cooldownBetweenAttacks);
+                                return true;
+                            }
                         }
                     }
                     return false;