+++ common/src/main/java/me/rufia/fightorflight/mixin/ActivityConfigurationContextMixin.java
@@ -4,9 +4,13 @@ import com.cobblemon.mod.common.api.ai.ActivityConfigurationContext;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import com.mojang.datafixers.util.Pair;
 import me.rufia.fightorflight.CobblemonFightOrFlight;
+import me.rufia.fightorflight.entity.ai.tasks.FOFBackUpIfTooClose;
 import me.rufia.fightorflight.entity.ai.tasks.FOFPokemonRangeTask;
+import me.rufia.fightorflight.utils.PokemonUtils;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.behavior.BackUpIfTooClose;
 import net.minecraft.world.entity.ai.behavior.BehaviorControl;
+import net.minecraft.world.entity.ai.behavior.declarative.BehaviorBuilder;
 import net.minecraft.world.entity.schedule.Activity;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -27,19 +31,20 @@ public abstract class ActivityConfigurationContextMixin {
     @Shadow(remap = false)
     private Activity activity;
 
-    @Inject(method = "apply", at = @At("HEAD"),remap = false)
+    @Inject(method = "apply", at = @At("HEAD"), remap = false)
     private void applyMixin(LivingEntity entity, CallbackInfo ci) {
         if (entity instanceof PokemonEntity) {
             if (Objects.equals(activity.getName(), "fight")) {
-                CobblemonFightOrFlight.LOGGER.info("{} get a range task.",entity.getName().getString());
-                /*
-                for(var t :tasks){
-                    CobblemonFightOrFlight.LOGGER.info("{} fight tasks:{}",entity.getName().getString(),t.getSecond().getClass().getName());
-                }
-                */
-                CobblemonFightOrFlight.LOGGER.info("{} PRINT END",entity.getName().getString());
-                tasks.add(new Pair<>(0, new FOFPokemonRangeTask()));//It can be added, but it doesn't seem to work.
-
+                tasks.add(new Pair<>(0, BehaviorBuilder.triggerIf(
+                        livingEntity -> {
+                            if (livingEntity instanceof PokemonEntity pokemonEntity) {
+                                return PokemonUtils.shouldShoot(pokemonEntity);
+                            }
+                            return false;
+                        },
+                        FOFBackUpIfTooClose.create(5, 0.75f)
+                )));
+                tasks.add(new Pair<>(1, new FOFPokemonRangeTask()));
             }
         }
     }